requirements.txt
pandas
numpy
matplotlib
python-dateutil
requests
PublicDataReader
openpyxl
jupyter 